Added a configurable soak test runner
