+++ src/Network/FeatherF7SoakTestRunner/FeatherF7SoakTestRunner.csproj
@@ -22,4 +22,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SoakTests\SoakTests.csproj" />
+  </ItemGroup>
 </Project>
