+++ src/Network/FeatherF7SoakTestRunner/FeatherF7SoakTestRunner.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FeatherF7Template", "FeatherF7Template.csproj", "{99980AEC-E7B0-432D-9763-A7928B3C742E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FeatherF7SoakTestRunner", "FeatherF7SoakTestRunner.csproj", "{99980AEC-E7B0-432D-9763-A7928B3C742E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
