+++ src/Network/FeatherF7SoakTestRunner/build.sh
@@ -2,4 +2,4 @@
 
 scriptdir="$( cd "$(dirname "$0")" ; pwd -P )"
 
-dotnet build FeatherF7Template.sln
+dotnet build FeatherF7SoakTestRunner.sln
