+++ src/Network/FeatherF7SoakTestRunner/meadow.config.yaml
@@ -5,24 +5,24 @@ Device:
     #
     #   Name of the device on the network.
     #
-    Name: HttpGetSoakTest
+    Name: FeatherF7Template
     #
     #   Is there an SD Card socket present on this device.
     #
-    SdCardPresent: true
+    # SdCardPresent: true
     #
     #   Pins reserved for OS use.
     #
-    ReservedPins: I9;H13
+    # ReservedPins: I9;H13
 #
 # Control how mono will start and operate.
 #
-# MonoControl:
+MonoControl:
     #
     #   Options to pass to Mono via the command line.
     #
     # Options: --debug --trace=N:System.Net.Http.ConnectHelper
-    # Options: --debug
+    Options: --debug
 #
 #   Network configuration.
 #
