--- src/Network/HttpGetSoakTest/wifi.config.yaml
@@ -1,3 +1,0 @@
-﻿Credentials:
-    Ssid: AccessPoint
-    Password: AccessPointPassword
