+++ src/Network/SoakTests/RegisteredTests.cs
@@ -1,0 +1,23 @@
+using System;
+namespace SoakTests;
+
+public static class RegisteredTests
+{
+    public static ISoakTest[] Tests = new ISoakTest[]
+    {
+        new HttpGetSoakTest()
+    };
+
+    public static ISoakTest GetTest(string testName)
+    {
+        foreach (var test in Tests)
+        {
+            if (test.Name == testName)
+            {
+                return test;
+            }
+        }
+
+        return null;
+    }
+}
